requirements.txt
django
djangorestframework
langchain
langchain-google-genai
python-dotenv
uvicorn